backend/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
websockets==12.0
ccxt==4.2.14
pandas==2.2.0
python-dotenv==1.0.1
openai==1.12.0
httpx==0.27.2
